Move attachments stuff somewhere more sane
