+++ src/components/transactions/TransactionView.vue
@@ -3,11 +3,11 @@ import type { Tag as TagObject, TagRecordParams } from "../../model/Tag";
 import type { Attachment } from "../../model/Attachment";
 import type { Location } from "../../model/Location";
 import type { Transaction } from "../../model/Transaction";
-import ConfirmDestroyFile from "../attachments/ConfirmDestroyFile.vue";
+import ConfirmDestroyFile from "../../pages/attachments/ConfirmDestroyFile.vue";
 import EditButton from "../buttons/EditButton.vue";
-import FileInput from "../attachments/FileInput.vue";
-import FileListItem from "../attachments/FileListItem.vue";
-import FileReattach from "../attachments/FileReattach.vue";
+import FileInput from "../../pages/attachments/FileInput.vue";
+import FileListItem from "../../pages/attachments/FileListItem.vue";
+import FileReattach from "../../pages/attachments/FileReattach.vue";
 import List from "../List.vue";
 import LocationIcon from "../../icons/Location.vue";
 import LocationView from "../locations/LocationView.vue";
