+++ src/pages/attachments/Attachments.vue
@@ -2,7 +2,7 @@
 import type { Attachment } from "../../model/Attachment";
 import ConfirmDestroyFile from "./ConfirmDestroyFile.vue";
 import FileListItem from "./FileListItem.vue";
-import List from "../List.vue";
+import List from "../../components/List.vue";
 import { ref, computed } from "vue";
 import { useAttachmentsStore, useUiStore } from "../../store";
