+++ src/pages/attachments/ConfirmDestroyFile.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import type { PropType } from "vue";
 import type { Attachment } from "../../model/Attachment";
-import ActionButton from "../buttons/ActionButton.vue";
-import Confirm from "../Confirm.vue";
+import ActionButton from "../../components/buttons/ActionButton.vue";
+import Confirm from "../../components/Confirm.vue";
 import { toRefs } from "vue";
 
 const emit = defineEmits(["yes", "no"]);
